Replace `circe` by `jsoniter-scala`
Circe has served us well and it's a great project. However, jsoniter is
better for bloop's needs because it's really fast and it has minimal
dependencies which means our build tool integrations are lighter.

Jsoniter's APIs are also great and they allow for a lot of customized
encoding/decoding that we might find useful in the future.

The most important part of this migration is the fact that Jsoniter's
defaults didn't match Circe's encoders of a the `mainClass` field.
Because of a bug in a custom encoder of circe, we were representing
`mainClass: Option[String]` with an array at the JSON level.

The new jsoniter encoder/decoder for `mainClass` supports now the
previous circe encoding + the most idiomatic encoding that it's the
default in jsoniter.

The speed that jsoniter brings us is important because the json
configuration files take a lot of space. It's paramount that we can
read/write them as fast as possible. This is especially important if we
want to write configuration files fast from the sbt plugins.