Add `tags` to Bloop configuration
This field is used to add tags to the build target described by a given
configuration file. The tags of this target will then be exposed via the
`tags` property of BSP.

Currently, the existing integrations set the `test` tag for test
targets, and the `library` tag for all other targets.