Fixes retrieval of `ModuleComponentIdentifier`
For some reason, the Bloop plugin wouldn't work at all in a private
repository using a proprietary version of Gradle. I believe this is
because the presence of a certain plugin that was interfering and
changing the artifacts available after resolution to be something else
than `ModuleComponentArtifactIdentifier`, so the Scala plugin would
never collect the Scala library. But, the result was that the Bloop
plugin wouldn't generate Bloop files and it would fail altogether.

The following change, which is untested because I couldn't reproduce,
has a fix that uses a more general interface `ModuleComponentIdentifier`
and pattern matches on `getComponentIdentifier` instead of the regular
artifact identifiers. This seems to be the most correct and general way
of writing the previous logic.