Adds a Bloop-specific Gradle project configuration
Some Gradle experts at $WORK have told me that the way we were
previously resolving artifacts wasn't strictly correct due to two reason:

- Project references in project configurations (like the ones we resolve
  in `getConfigurationArtifacts`) are allowed to be empty for now, but
  that will change, and it's a bug. https://github.com/gradle/gradle/issues/26155

- The way we currently resolve artifacts doesn't let Gradle know our
  intention (as we can resolve any arbitrary configuration) and it
  doesn't play well with Gradle transforms, which require all of the
  resolved configuration artifacts to be present, otherwise it fails
  with a `FileNotFoundException`.

As a result, I changed the code to introduce a Gradle configuration
`bloopConfig` that extends all those configurations we care about, and
change the resolution code to only depend on that configuration.

It seems that solution works and passes all tests, and it avoids the
`FileNotFoundException` error I found out.

Because we now have a proper `configuration`, it means that users that
create their own configuration and want it to be exported to bloop need
to do `bloopConfig.extendsFrom(myConfig)` where `myConfig` is a
user-defined configuration. The reason for this is because when we
declare that our configuration extends all the other existing
configuration, we do it at evaluation time. User-defined configurations
run after the bloop plugin, so they never get added unless the user
wishes to. This is not a big deal as custom configuration are extremely
rare, and this is a common practice in Gradle land when you define your
own configuration.